auth_project/requirements.txt
Django>=3.0,<4.0
djangorestframework
djangorestframework-simplejwt